finished gameApi

+++ backend_typescript/models/Game.js
@@ -87,27 +87,48 @@ class Game {
             return new Result_1.Failure('Server error', 500);
         }
     }
+    verifyAcceptDecline(reqUserId) {
+        if (reqUserId !== this.user2) {
+            return 'You are not the correct opponent player';
+        }
+        if (this.state !== 'invited') {
+            return 'Cannot accept or delete an active game';
+        }
+        return 'Action can be made';
+    }
     async updateGameState(state) {
         this.state = state;
         const game = await db.updateGameState(this.id, this.state);
         return new Game(game.id, game.user1, game.user2, game.dimension, game.state);
     }
-    async deleteGame() {
-        try {
-            await db.deleteGame(this.id);
-            return new Result_1.Success('Game deleted');
+    async deleteGame(reqUserId) {
+        const result = this.verifyAcceptDecline(reqUserId);
+        if (result === 'Action can be made') {
+            try {
+                await db.deleteGame(this.id);
+                return new Result_1.Success('Game deleted');
+            }
+            catch (e) {
+                return new Result_1.Failure('Server error', 500);
+            }
         }
-        catch (e) {
-            return new Result_1.Failure('Server error', 500);
+        else {
+            return new Result_1.Failure(result, 401);
         }
     }
-    async acceptGame() {
-        try {
-            const game = await this.updateGameState('accepted');
-            return new Result_1.Success(game);
+    async acceptGame(reqUserId) {
+        const result = this.verifyAcceptDecline(reqUserId);
+        if (result === 'Action can be made') {
+            try {
+                const game = await this.updateGameState('accepted');
+                return new Result_1.Success(game);
+            }
+            catch (e) {
+                return new Result_1.Failure('Server error', 500);
+            }
         }
-        catch (e) {
-            return new Result_1.Failure('Server error', 500);
+        else {
+            return new Result_1.Failure(result, 401);
         }
     }
     async userIsReady(reqUserId) {
@@ -156,7 +177,7 @@ class Game {
         }
         ;
         const hitResults = await db.getHitsResults(this.id, opponent, row, col);
-        if (Number(hitResults) === 1) {
+        if (Number(hitResults.hits) === 1) {
             const timestamp = new Date(Date.now());
             await db.insertIntoShots(this.id, player, row, col, true, timestamp);
             const successfullShots = userGameShots.filter(shot => shot.hit === true);
